onion: Hack in support for Onion Omega2+
The Omega2+ ships with a U-Boot that doesn't have support for TFTP and
the source hasn't been released...

In order to load kernels over TFTP, we compile this U-Boot and save it on a USB
stick. This U-Boot is then loaded over USB and used to subsequently TFTP the
kernel.

This is a pretty gross hack...

In order to get the Onion Omega2+ working:
    - Change from UART2 to UART0.
    - Change Baud rate from 57600 to 115200.
    - Remove code to check for GPIO button press.

Signed-off-by: Harvey Hunt <harvey.hunt@imgtec.com>